requirements.txt
numpy==1.19.5
pdf2image==1.15.1
Pillow==8.2.0
pkg-resources==0.0.0
